+++ lightly_studio/src/lightly_studio/resolvers/annotation_resolver/get_all_by_collection_id_and_parent_sample_ids.py
@@ -8,6 +8,7 @@ from uuid import UUID
 from sqlalchemy.orm import joinedload
 from sqlmodel import Session, col, select
 
+from lightly_studio import db_array
 from lightly_studio.models.annotation.annotation_base import AnnotationBaseTable, AnnotationType
 from lightly_studio.models.sample import SampleTable
 
@@ -33,13 +34,19 @@ def get_all_by_collection_id_and_parent_sample_ids(
             samples must belong to.
         annotation_type: Annotation type to filter by.
     """
+    if not parent_sample_ids:
+        return []
     statement = (
         select(AnnotationBaseTable)
         .join(
             SampleTable,
             col(SampleTable.sample_id) == col(AnnotationBaseTable.sample_id),
         )
-        .where(col(AnnotationBaseTable.parent_sample_id).in_(parent_sample_ids))
+        .where(
+            db_array.in_array(
+                column=col(AnnotationBaseTable.parent_sample_id), values=parent_sample_ids
+            )
+        )
         .where(col(AnnotationBaseTable.annotation_type) == annotation_type)
         .where(col(SampleTable.collection_id) == annotation_collection_id)
     )